requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
idna==2.5
infi.clickhouse-orm==0.7.2
lxml==3.7.3
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
requests==2.16.0
six==1.10.0
urllib3==1.21.1
